backend/requirements.txt
# FastAPI e dependências
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Caching
redis==5.0.1
aioredis==2.0.1

# Celery
celery==5.3.6
flower==2.0.1

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Security
python-dotenv==1.0.0
cryptography==42.0.0
bcrypt==4.1.2

# HTTP Clients
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# WebSocket
websockets==12.0

# File Processing
openpyxl==3.1.2
pandas==2.2.0
python-magic==0.4.27
Pillow==10.2.0
PyPDF2==3.0.1

# Payment Gateways
stripe==8.2.0
mercadopago==2.2.1

# Email
#python-email==0.1.1
aiosmtplib==3.0.1
jinja2==3.1.3

# Utilities
python-dateutil==2.8.2
pytz==2024.1
phonenumbers==8.13.27
validators==0.22.0

# Monitoring & Logging
sentry-sdk[fastapi]==1.40.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0
faker==22.6.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.6

# Utils
python-slugify==8.0.2
shortuuid==1.0.11

asyncpg
psycopg2-binary
asyncpg
asyncpg
asyncpg
asyncpg
